backend-fastapi-tenants/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy==2.0.36
asyncpg==0.29.0
pydantic==2.9.2
aiokafka==0.10.0
python-jose==3.3.0
kafka-python